+++ packages/cli/src/commands/codegen/index.ts
@@ -4,7 +4,7 @@
 import {Command, Flags} from '@oclif/core';
 import {getProjectRootAndManifest, getSchemaPath} from '@subql/common';
 import {codegen} from '../../controller/codegen-controller';
-import {resolveToAbsolutePath} from '../../utils';
+import {resolveToAbsolutePath, buildManifestFromLocation, checkForTsManifest} from '../../utils';
 
 export default class Codegen extends Command {
   static description = 'Generate schemas for graph node';
@@ -26,6 +26,10 @@ export default class Codegen extends Command {
     const {file, location} = flags;
 
     const projectPath = resolveToAbsolutePath(file ?? location ?? process.cwd());
+
+    if (checkForTsManifest(projectPath)) {
+      await buildManifestFromLocation(projectPath, this);
+    }
 
     const {manifests, root} = getProjectRootAndManifest(projectPath);