+++ packages/common/src/project/load.ts
@@ -89,7 +89,7 @@ export function getFileContent(path: string, identifier: string): string {
 //  Validate generic/common section for project manifest
 export function validateCommonProjectManifest(raw: unknown): void {
   const projectManifest = plainToClass<CommonProjectManifestV1_0_0Impl, unknown>(CommonProjectManifestV1_0_0Impl, raw);
-  const errors = validateSync(projectManifest, {whitelist: true, forbidNonWhitelisted: true});
+  const errors = validateSync(projectManifest, {whitelist: true});
   if (errors?.length) {
     // TODO: print error details
     const errorMsgs = errors.map((e) => e.toString()).join('\n');