+++ packages/node-core/src/indexer/poi/index.ts
@@ -3,4 +3,5 @@
 
 export * from './poiModel';
 export * from './poi.service';
+export * from './poiSync.service';
 export * from './PoiBlock';